chore(vuex,names): rename mutation-types and action-types
