+++ src/renderer/store/actionTypes.js
